+++ src/cjs/CubicBezierCurve.js
@@ -21,7 +21,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `toSVGPathData` function (deprecated). Use `drawutilssvg` instead.
- * @version 2.6.0
+ * @modified 2022-10-17 Addint these method from the `PathSegment` interface: revert.
+ * @version 2.7.0
  *
  * @file CubicBezierCurve
  * @public
@@ -73,7 +74,7 @@ var CubicBezierCurve = /** @class */ (function () {
         // An array of floats
         this.segmentLengths = [];
         // float
-        this.arcLength = null;
+        // this.arcLength = null;
         this.updateArcLengths();
     }
     /**
@@ -504,6 +505,17 @@ var CubicBezierCurve = /** @class */ (function () {
      **/
     CubicBezierCurve.prototype.clone = function () {
         return new CubicBezierCurve(this.getStartPoint().clone(), this.getEndPoint().clone(), this.getStartControlPoint().clone(), this.getEndControlPoint().clone());
+    };
+    CubicBezierCurve.prototype.revert = function () {
+        var sp = this.startPoint;
+        var scp = this.startControlPoint;
+        var ep = this.endPoint;
+        var ecp = this.endControlPoint;
+        this.startPoint = ep;
+        this.startControlPoint = ecp;
+        this.endPoint = sp;
+        this.endControlPoint = scp;
+        return this;
     };
     /**
      * Check if this and the specified curve are equal.<br>