+++ src/esm/CubicBezierCurve.js
@@ -20,7 +20,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `toSVGPathData` function (deprecated). Use `drawutilssvg` instead.
- * @version 2.6.0
+ * @modified 2022-10-17 Addint these method from the `PathSegment` interface: revert.
+ * @version 2.7.0
  *
  * @file CubicBezierCurve
  * @public
@@ -70,7 +71,7 @@ export class CubicBezierCurve {
         // An array of floats
         this.segmentLengths = [];
         // float
-        this.arcLength = null;
+        // this.arcLength = null;
         this.updateArcLengths();
     }
     /**
@@ -501,6 +502,17 @@ export class CubicBezierCurve {
      **/
     clone() {
         return new CubicBezierCurve(this.getStartPoint().clone(), this.getEndPoint().clone(), this.getStartControlPoint().clone(), this.getEndControlPoint().clone());
+    }
+    revert() {
+        const sp = this.startPoint;
+        const scp = this.startControlPoint;
+        const ep = this.endPoint;
+        const ecp = this.endControlPoint;
+        this.startPoint = ep;
+        this.startControlPoint = ecp;
+        this.endPoint = sp;
+        this.endControlPoint = scp;
+        return this;
     }
     /**
      * Check if this and the specified curve are equal.<br>