+++ src/esm/interfaces/index.d.ts
@@ -1,3 +1,4 @@
 export * from "./core";
 export * from "./DrawLib";
 export * from "./externals";
+export * from "./additionals";