+++ src/ts/CubicBezierCurve.ts
@@ -20,7 +20,8 @@
  * @modified 2021-01-20 Added UID.
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `toSVGPathData` function (deprecated). Use `drawutilssvg` instead.
- * @version 2.6.0
+ * @modified 2022-10-17 Addint these method from the `PathSegment` interface: revert.
+ * @version 2.7.0
  *
  * @file CubicBezierCurve
  * @public
@@ -30,7 +31,7 @@ import { Bounds } from "./Bounds";
 import { UIDGenerator } from "./UIDGenerator";
 import { Vertex } from "./Vertex";
 import { Vector } from "./Vector";
-import { XYCoords, UID } from "./interfaces";
+import { XYCoords, UID, PathSegment } from "./interfaces";
 
 /**
  * @classdesc A refactored cubic bezier curve class.
@@ -42,7 +43,7 @@ import { XYCoords, UID } from "./interfaces";
  * @requires UID
  * @requires UIDGenerator
  */
-export class CubicBezierCurve {
+export class CubicBezierCurve implements PathSegment {
   /** @constant {number} */
   static readonly START_POINT: number = 0;
   /** @constant {number} */
@@ -157,7 +158,7 @@ export class CubicBezierCurve {
     // An array of floats
     this.segmentLengths = [];
     // float
-    this.arcLength = null;
+    // this.arcLength = null;
 
     this.updateArcLengths();
   }
@@ -625,6 +626,18 @@ export class CubicBezierCurve {
       this.getStartControlPoint().clone(),
       this.getEndControlPoint().clone()
     );
+  }
+
+  revert(): CubicBezierCurve {
+    const sp = this.startPoint;
+    const scp = this.startControlPoint;
+    const ep = this.endPoint;
+    const ecp = this.endControlPoint;
+    this.startPoint = ep;
+    this.startControlPoint = ecp;
+    this.endPoint = sp;
+    this.endControlPoint = scp;
+    return this;
   }
 
   /**