+++ src/ts/utils/datastructures/GirihTile.ts
@@ -319,7 +319,7 @@ export abstract class GirihTile extends Polygon {
    * @param {Polygon} neighbourTile - The polygon (or tile) you want to find adjacency for at the specified edge.
    * @return {Polygon|null} the passed tile itself if adjacency was found, null otherwise.
    */
-  transformTilePositionToAdjacency<P extends Polygon>(baseEdgeIndex: number, neighbourTile: P): P {
+  transformTilePositionToAdjacency<P extends Polygon>(baseEdgeIndex: number, neighbourTile: P): P | null {
     // Find the position for that tile to match (might not exist)
     // { edgeIndex:number, offset:XYCoords }
     var adjacency = this.findAdjacentTilePosition(baseEdgeIndex, neighbourTile);